+++ app/jobs/github_repo_fetcher_job.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class GithubRepoFetcherJob < ApplicationJob
+class GitHubRepoFetcherJob < ApplicationJob
   queue_as :http_request
 
   def perform(user_id)
