+++ app/models/comment.rb
@@ -1,9 +1,7 @@
 # frozen_string_literal: true
 
 class Comment < ApplicationRecord
-  include MarkdownBody
-  include Mentionable
-  include UserAvatarDelegate
+  include Markdownable, Mentionable, UserAvatarDelegate
 
   belongs_to :commentable, polymorphic: true
   belongs_to :user, optional: true
