+++ app/models/concerns/mention_topic.rb
@@ -11,19 +11,6 @@ module MentionTopic
     after_save :create_releated_for_mentioned_topics
   end
 
-  def create_releated_for_mentioned_topics
-    extract_mentioned_topic_ids
-    return false if self.mentioned_topic_ids.blank?
-    Reply.transaction do
-      self.mentioned_topic_ids.each do |topic_id|
-        topic = Topic.find_by(id: topic_id)
-        next if topic.blank?
-        next if topic.replies.where(target: self).any?
-        Reply.create_system_event!(user: self.user, topic: topic, action: "mention", target: self)
-      end
-    end
-  end
-
   def extract_mentioned_topic_ids
     matched_ids = body.scan(TOPIC_LINK_REGEXP).flatten
     current_topic_id = self.class.name == "Topic" ? self.id : self.topic_id
@@ -32,4 +19,18 @@ module MentionTopic
       self.mentioned_topic_ids = Topic.where("id IN (?)", matched_ids).pluck(:id)
     end
   end
+
+  private
+    def create_releated_for_mentioned_topics
+      extract_mentioned_topic_ids
+      return false if self.mentioned_topic_ids.blank?
+      Reply.transaction do
+        self.mentioned_topic_ids.each do |topic_id|
+          topic = Topic.find_by(id: topic_id)
+          next if topic.blank?
+          next if topic.replies.where(target: self).any?
+          Reply.create_system_event!(user: self.user, topic: topic, action: "mention", target: self)
+        end
+      end
+    end
 end
