--- app/models/user/topic_favorite.rb
@@ -1,15 +1,0 @@
-# frozen_string_literal: true
-
-class User
-  module TopicFavorite
-    extend ActiveSupport::Concern
-
-    included do
-      action_store :favorite, :topic
-    end
-
-    def favorites_count
-      favorite_topic_actions.count
-    end
-  end
-end
